--- src/main/java/Model_Entitiy/TableEntity.java
@@ -1,4 +1,0 @@
-package Model_Entitiy;
-
-public class TableEntity {
-}
