--- src/main/java/Repository/RoomCategoryRepositry.java
@@ -1,4 +1,0 @@
-package Repository;
-
-public interface RoomCategoryRepositry {
-}
