--- src/main/java/Repository/RoomRepositry.java
@@ -1,4 +1,0 @@
-package Repository;
-
-public interface RoomRepositry {
-}
