+++ src/main/java/com/example/PlaceAdminister/Controller/RoomCategoryController.java
@@ -1,7 +1,7 @@
-package Controller;
+package com.example.PlaceAdminister.Controller;
 
-import Model_Entitiy.RoomCategoryEntity;
-import Service.RoomCategoryService;
+import com.example.PlaceAdminister.Model_Entitiy.RoomCategoryEntity;
+import com.example.PlaceAdminister.Service.RoomCategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
