+++ src/main/java/com/example/PlaceAdminister/Model_Entitiy/TableEntity.java
@@ -1,0 +1,4 @@
+package com.example.PlaceAdminister.Model_Entitiy;
+
+public class TableEntity {
+}
