+++ src/main/java/com/example/PlaceAdminister/Repository/RoomCategoryRepositry.java
@@ -1,0 +1,4 @@
+package com.example.PlaceAdminister.Repository;
+
+public interface RoomCategoryRepositry {
+}
