+++ src/main/java/com/example/PlaceAdminister/Service/RoomService.java
@@ -1,12 +1,13 @@
-package Service;
+package com.example.PlaceAdminister.Service;
 
-import Model_Entitiy.RoomEntity;
+import com.example.PlaceAdminister.Model_Entitiy.RoomEntity;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.stereotype.Service;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 
@@ -28,15 +29,13 @@ public class RoomService {
             Integer age
     ){
     }
-    // Other methods as needed
     public List<RoomEntity> readFromJsonFile(String filePath) {
         try {
             ObjectMapper objectMapper = new ObjectMapper();
             List<RoomEntity> models = objectMapper.readValue(new File(filePath), new TypeReference<>() {});
             return models;
         } catch (IOException e) {
-            e.printStackTrace(); // Handle the exception appropriately in a production environment
-            return null;
+            return new ArrayList<>();
         }
     }
     public void writeToJsonFile(List<RoomEntity> models, String filePath) {
