+++ src/main/resources/Rooms.json
@@ -1,0 +1,1 @@
+[{"categories":[],"id":null,"max_num_of_chairs":20,"categoriesId":[],"status":false,"time_0f_reservation":null},{"categories":[],"id":null,"max_num_of_chairs":20,"categoriesId":[],"status":false,"time_0f_reservation":null},{"categories":[],"id":null,"max_num_of_chairs":20,"categoriesId":[],"status":false,"time_0f_reservation":null}]
