--- src/main/resources/application.properties
@@ -1,2 +1,0 @@
-server.servlet.context-path=/your-context-path
-spring.main.web-application-type= servlet
